LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.fontKerning-expected.txt

2d.text.drawing.style.fontKerning
Testing basic functionalities of fontKerning for canvas
Actual output:

FAIL Testing basic functionalities of fontKerning for canvas assert_equals: ctx.fontKerning === "auto" (got [undefined], expected auto[string]) expected (string) "auto" but got (undefined) undefined
